Conflicts:
	include/ARX_Common.h
	src/danae/ARX_Menu2.cpp

+++ src/danae/ARX_Minimap.cpp
@@ -204,7 +204,7 @@ void ARX_MINIMAP_Load_Offsets()
 	{
 		size_t siz = 0;
 		char * dat = (char *)PAK_FileLoadMalloc(INI_MINI_OFFSETS, &siz);
-		long pos = 0;
+		size_t pos = 0;
 
 		if (dat)
 		{
@@ -538,7 +538,7 @@ void ARX_MINIMAP_Show(LPDIRECT3DDEVICE7 m_pd3dDevice, long SHOWLEVEL, long flag,
 							F2L((float)(vv * 255.f), &r);
 
 
-							long ucLevel =  max(r, minimap[SHOWLEVEL].revealed[i][j]);
+							long ucLevel =  max(r, (long)minimap[SHOWLEVEL].revealed[i][j]);
 							ARX_CHECK_UCHAR(ucLevel);
 
 							minimap[SHOWLEVEL].revealed[i][j] = ARX_CLEAN_WARN_CAST_UCHAR(ucLevel);
@@ -605,7 +605,7 @@ void ARX_MINIMAP_Show(LPDIRECT3DDEVICE7 m_pd3dDevice, long SHOWLEVEL, long flag,
 						oo += v;
 
 						if ((i + 1 < 0) || (i + 1 >= MINIMAP_MAX_X) || (j < 0) || (j >= MINIMAP_MAX_Z)) v = 0;
-						else v = ((float)minimap[SHOWLEVEL].revealed[min(i+1, MINIMAP_MAX_X-1)][j]) * DIV255;
+						else v = ((float)minimap[SHOWLEVEL].revealed[min(i+1, (long)MINIMAP_MAX_X-1)][j]) * DIV255;
 
 						if (flag == 1)
 						{
@@ -638,7 +638,7 @@ void ARX_MINIMAP_Show(LPDIRECT3DDEVICE7 m_pd3dDevice, long SHOWLEVEL, long flag,
 						oo += v;
 
 						if ((i + 1 < 0) || (i + 1 >= MINIMAP_MAX_X) || (j + 1 < 0) || (j + 1 >= MINIMAP_MAX_Z)) v = 0;
-						else v = ((float)minimap[SHOWLEVEL].revealed[min(i+1, MINIMAP_MAX_X-1)][min(j+1, MINIMAP_MAX_Z-1)]) * DIV255;
+						else v = ((float)minimap[SHOWLEVEL].revealed[min(i+1, (long)MINIMAP_MAX_X-1)][min(j+1, (long)MINIMAP_MAX_Z-1)]) * DIV255;
 
 						if (flag == 1)
 						{
@@ -672,7 +672,7 @@ void ARX_MINIMAP_Show(LPDIRECT3DDEVICE7 m_pd3dDevice, long SHOWLEVEL, long flag,
 						oo += v;
 
 						if ((i < 0) || (i >= MINIMAP_MAX_X) || (j + 1 < 0) || (j + 1 >= MINIMAP_MAX_Z)) v = 0;
-						else v = ((float)minimap[SHOWLEVEL].revealed[i][min(j+1, MINIMAP_MAX_Z-1)]) * DIV255;
+						else v = ((float)minimap[SHOWLEVEL].revealed[i][min(j+1, (long)MINIMAP_MAX_Z-1)]) * DIV255;
 
 						if (flag == 1)
 						{
